séparation service-repo ok

+++ src/Service/SiteService.php
@@ -2,8 +2,10 @@
 
 namespace App\Service;
 
+use App\Entity\Site;
 use App\Repository\SiteRepository;
-use Doctrine\ORM\EntityManagerInterface;
+use App\ServiceResult\Site\DeleteSiteResult;
+use App\ServiceResult\Site\UpdateSiteResult;
 
 class SiteService
 {
@@ -22,24 +24,55 @@ class SiteService
     /**
      * Crée un nouveau site
      */
-    public function create(string $nomSite): Site
+//    public function create(string $nomSite): Site
+//    {
+//        $site = new Site();
+//        $site->setNomSite($nomSite);
+//
+//        $this->em->persist($site);
+//        $this->em->flush();
+//
+//        return $site;
+//    }
+    public function createSite(string $nomSite): Site
     {
         $site = new Site();
         $site->setNomSite($nomSite);
 
-        $this->em->persist($site);
-        $this->em->flush();
+        $this->siteRepository->save($site);
 
         return $site;
     }
 
+
     /**
      * Supprime un site
      */
-    public function delete(Site $site): void
+    public function deleteSite(Site $site): DeleteSiteResult
     {
-        //Logique métier : on ne supprime pas si des activités utilisent ce site
+        // Vérifie si des participants sont liés à ce site
+        if (count($site->getParticipants()) > 0) {
+            return DeleteSiteResult::SITE_IN_USE;
+        }
+
+        // (Tu peux ajouter d’autres checks ici, ex: sorties associées)
         $this->siteRepository->remove($site);
+
+        return DeleteSiteResult::SUCCESS;
+    }
+
+    public function updateSite(Site $site, string $newNom): UpdateSiteResult
+    {
+//        $existing = $this->siteRepository->findOneBy(['nom_site' => $newNom]);
+//
+//        if ($existing && $existing->getId() !== $site->getId()) {
+//            return UpdateSiteResult::NAME_ALREADY_USED;
+//        }
+
+        $site->setNomSite($newNom);
+        $this->siteRepository->save($site);
+
+        return UpdateSiteResult::SUCCESS;
     }
 
     /**
